+++ tests/paste-middle/change.js
@@ -1,0 +1,14 @@
+
+module.exports = function(plugin, change) {
+    return plugin.onPaste(
+        {
+            preventDefault() {},
+            stopPropagation() {}
+        },
+        {
+            type: 'text',
+            text: 'Yes\nNo\nQuestion?'
+        },
+        change
+    );
+};