+++ tests/shift-tab-middle-offset/change.js
@@ -1,0 +1,11 @@
+
+module.exports = function(plugin, change) {
+    return plugin.onKeyDown(
+        {
+            preventDefault() {},
+            stopPropagation() {}
+        },
+        { key: 'tab', isShift: true },
+        change
+    );
+};