+++ tests/tab-start-offset/change.js
@@ -1,0 +1,18 @@
+
+module.exports = function(plugin, change) {
+    const { state } = change;
+    const block = state.document.findDescendant(node => node.type == 'code_block');
+
+    change
+        .collapseToStartOf(block)
+        .moveOffsetsTo(0);
+
+    return plugin.onKeyDown(
+        {
+            preventDefault() {},
+            stopPropagation() {}
+        },
+        { key: 'tab' },
+        change
+    );
+};