+++ tests/togglecodeblock-code/change.js
@@ -1,0 +1,7 @@
+
+module.exports = function(plugin, change) {
+    return plugin.changes.toggleCodeBlock(
+        change,
+        'paragraph'
+    );
+};