+++ tests/togglecodeblock-code/transform.js
@@ -1,8 +1,0 @@
-
-module.exports = function(plugin, state) {
-    return plugin.transforms.toggleCodeBlock(
-        state.transform(),
-        'paragraph'
-    )
-    .apply();
-};