+++ tests/unwrapcodeblock-multi-lines/change.js
@@ -1,0 +1,7 @@
+
+module.exports = function(plugin, change) {
+    return plugin.changes.unwrapCodeBlock(
+        change,
+        'paragraph'
+    );
+};