+++ tests/unwrapcodeblock-selection/change.js
@@ -1,0 +1,9 @@
+
+module.exports = function(plugin, change) {
+    const newState = plugin.changes.unwrapCodeBlock(
+        change,
+        'paragraph'
+    );
+
+    return newState;
+};