+++ tests/wrapcodeblock-normal/change.js
@@ -1,0 +1,4 @@
+
+module.exports = function(plugin, change) {
+    return plugin.changes.wrapCodeBlock(change);
+};