+++ tests/wrapcodeblock-with-inline/transform.js
@@ -1,7 +1,0 @@
-
-module.exports = function(plugin, state) {
-    return plugin.transforms.wrapCodeBlock(
-        state.transform()
-    )
-    .apply();
-};